reorganized commands
