+++ src/main/java/frc/robot/commands/algae/ShootAlgaeCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.algae;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.algae.AlgaeSubsystem;
