+++ src/main/java/frc/robot/commands/coral/CoralWristFollowCommand.java
@@ -1,15 +1,15 @@
-package frc.robot.commands;
+package frc.robot.commands.coral;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.subsystems.coral.CoralSubsystem;
 
-public class WristFollowCommand extends Command {
+public class CoralWristFollowCommand extends Command {
    
     private final CoralSubsystem subsystem;
     private final CommandXboxController operator;
     
-    public WristFollowCommand(CoralSubsystem subsystem, CommandXboxController operator) {
+    public CoralWristFollowCommand(CoralSubsystem subsystem, CommandXboxController operator) {
         this.subsystem = subsystem;
         this.operator = operator;
